requirements.txt
annotated-types==0.7.0
anyio==4.9.0
azure-ai-documentintelligence==1.0.2
azure-ai-textanalytics==5.3.0
azure-common==1.1.28
azure-core==1.34.0
azure-data-tables==12.7.0
azure-functions==1.23.0
azure-identity==1.23.0
azure-keyvault-secrets==4.10.0
azure-search-documents==11.5.2
azure-storage-blob==12.25.1
blinker==1.9.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.4
distro==1.9.0
fastapi==0.115.13
Flask==3.1.1
flask-cors==6.0.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
MarkupSafe==3.0.2
msal==1.32.3
msal-extensions==1.3.1
multidict==6.5.0
openai==1.88.0
propcache==0.3.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
PyMuPDF==1.26.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
requests==2.32.4
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
uvicorn==0.34.3
Werkzeug==3.1.3
yarl==1.20.1
